Refactor formulas into modules and centralize PS rates
