+++ constants.py
@@ -1,0 +1,2 @@
+PS_RATE_BEFORE_2018 = 0.155
+PS_RATE_AFTER_2018 = 0.172
